added userId as parameter to get and delete objects from the database

+++ database/queries/budgets.sql
@@ -3,12 +3,8 @@ INSERT INTO budgets (id, created_at, updated_at, amount, goal, start_date, end_d
 VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
 RETURNING *;
 
--- name: UpdateBudgetAmount :exec
-UPDATE budgets SET amount = amount + $2
-WHERE category_id = $1 AND start_date <= $3 AND end_date >= $3;
-
--- name: DeleteBudget :exec
-DELETE FROM budgets WHERE id = $1;
+-- name: DeleteBudget :one
+DELETE FROM budgets WHERE id = $1 AND user_id = $2 RETURNING *;
 
 -- name: GetUserBudgetsPaged :many
 SELECT * FROM budgets WHERE user_id = $1
@@ -22,4 +18,10 @@ ORDER BY created_at ASC, id DESC
 LIMIT $2;
 
 -- name: GetBudgetByID :one
-SELECT * FROM budgets WHERE id = $1;
+SELECT * FROM budgets WHERE id = $1 AND user_id = $2;
+
+-- name: UpdateBudgetAmount :exec
+-- Since UpdateBudgetAmount is only called by the API, there is no need to
+-- check if the user is the owner of the budget since the API already does that
+UPDATE budgets SET amount = amount + $2
+WHERE category_id = $1 AND start_date <= $3 AND end_date >= $3;
