+++ database/queries/categories.sql
@@ -3,8 +3,8 @@ INSERT INTO categories (id, created_at, updated_at, name, user_id)
 VALUES ($1, $2, $3, $4, $5)
 RETURNING *;
 
--- name: DeleteCategory :exec
-DELETE FROM categories WHERE id = $1;
+-- name: DeleteCategory :one
+DELETE FROM categories WHERE id = $1 AND user_id = $2 RETURNING *;
 
 -- name: GetUserCategoriesPaged :many
 SELECT * FROM categories WHERE user_id = $1
@@ -18,4 +18,4 @@ ORDER BY created_at ASC, id DESC
 LIMIT $2;
 
 -- name: GetCategoryByID :one
-SELECT * FROM categories WHERE id = $1;
+SELECT * FROM categories WHERE id = $1 AND user_id = $2;
